Read me
This program visits the website Books to Scrape and gets all the links for the categories of books and if the categories have more than one page it manages to get the links for the next pages too. After that, it gets the links for the personal page of every book in each category. The program extracts information for each book such as the title, the number of books available, description, etc. This information is put in lists to use later to create a DataFrame. Using the DataFrame the program creates a CSV file for each category of books. The program creates a folder with all the images of all the books in each category.